[tests] Move to new testing paradigm in Rust
