+++ .github/workflows/macos.yml
@@ -256,7 +256,7 @@ jobs:
         trap 'echo "MULTIPASS_TESTS_EXIT_CODE=$?" >> $GITHUB_ENV' EXIT
         # Set soft limit for the core file size (512MiB)
         ulimit -c 1048576
-        ctest
+        ctest --output-on-failure
 
     - name: Run integration tests
       working-directory: ${{ env.BUILD_DIR }}
