+++ rxx/petname/Cargo.toml
@@ -21,7 +21,7 @@ edition.workspace = true
 rust-version.workspace = true
 
 [lib]
-crate-type = ["staticlib"]
+crate-type = ["staticlib","rlib"]
 
 [features]
 default = []
